Rename file with correct spelling and change associated imports to reference corrected filename.
